move the QHotkeyApplication to windows only
